+++ packages/nest/src/implement.test.ts
@@ -1049,244 +1049,3 @@ describe('compatibility', () => {
     await app.close()
   })
 })
-
-describe('dynamic context factory', () => {
-  it('should work with a static context object (regression)', async () => {
-    const contract = oc
-      .meta(openapi({ path: '/static-ctx' }))
-      .output(z.object({ value: z.string() }))
-
-    let capturedContext: any = null
-
-    @Controller()
-    class ImplController {
-      @Implement(contract)
-      handler() {
-        return implement(contract).handler(({ context }) => {
-          capturedContext = context
-          return { value: 'ok' }
-        })
-      }
-    }
-
-    const moduleRef = await Test.createTestingModule({
-      controllers: [ImplController],
-      imports: [ORPCModule.forRoot({ context: { db: 'my-database' } as any })],
-    }).compile()
-
-    const app = moduleRef.createNestApplication()
-    await app.init()
-
-    const res = await supertest(app.getHttpServer()).post('/static-ctx')
-    expect(res.status).toBe(200)
-    expect(capturedContext).toEqual({ db: 'my-database' })
-
-    await app.close()
-  })
-
-  it('should support a synchronous context factory function', async () => {
-    const contract = oc
-      .meta(openapi({ path: '/sync-ctx' }))
-      .output(z.object({ method: z.string() }))
-
-    let capturedContext: any = null
-
-    @Controller()
-    class ImplController {
-      @Implement(contract)
-      handler() {
-        return implement(contract).handler(({ context }) => {
-          capturedContext = context
-          return { method: (context as any).method }
-        })
-      }
-    }
-
-    const moduleRef = await Test.createTestingModule({
-      controllers: [ImplController],
-      imports: [
-        ORPCModule.forRoot({
-          context: (ctx) => {
-            const req = ctx.switchToHttp().getRequest()
-            return { method: req.method } as any
-          },
-        }),
-      ],
-    }).compile()
-
-    const app = moduleRef.createNestApplication()
-    await app.init()
-
-    const res = await supertest(app.getHttpServer()).post('/sync-ctx')
-    expect(res.status).toBe(200)
-    expect(res.body).toEqual({ method: 'POST' })
-    expect((capturedContext as any).method).toBe('POST')
-
-    await app.close()
-  })
-
-  it('should support an async context factory function', async () => {
-    const contract = oc
-      .meta(openapi({ path: '/async-ctx' }))
-      .output(z.object({ userId: z.string() }))
-
-    let capturedContext: any = null
-
-    @Controller()
-    class ImplController {
-      @Implement(contract)
-      handler() {
-        return implement(contract).handler(({ context }) => {
-          capturedContext = context
-          return { userId: (context as any).userId }
-        })
-      }
-    }
-
-    const moduleRef = await Test.createTestingModule({
-      controllers: [ImplController],
-      imports: [
-        ORPCModule.forRoot({
-          context: async (_ctx) => {
-            await new Promise(resolve => setTimeout(resolve, 5))
-            return { userId: 'user-123' } as any
-          },
-        }),
-      ],
-    }).compile()
-
-    const app = moduleRef.createNestApplication()
-    await app.init()
-
-    const res = await supertest(app.getHttpServer()).post('/async-ctx')
-    expect(res.status).toBe(200)
-    expect(res.body).toEqual({ userId: 'user-123' })
-    expect(capturedContext).toEqual({ userId: 'user-123' })
-
-    await app.close()
-  })
-
-  it('should pass dynamic context to handler', async () => {
-    const contract = oc
-      .meta(openapi({ path: '/dynamic-ctx-handler' }))
-      .input(z.object({ name: z.string() }))
-      .output(z.object({ greeting: z.string(), userId: z.string() }))
-
-    @Controller()
-    class ImplController {
-      @Implement(contract)
-      handler() {
-        return implement(contract).handler(({ input, context }) => ({
-          greeting: `Hello, ${input.name}`,
-          userId: (context as any).userId,
-        }))
-      }
-    }
-
-    const moduleRef = await Test.createTestingModule({
-      controllers: [ImplController],
-      imports: [
-        ORPCModule.forRoot({
-          context: _ctx => ({ userId: 'dynamic-user-456' } as any),
-        }),
-      ],
-    }).compile()
-
-    const app = moduleRef.createNestApplication()
-    await app.init()
-
-    const res = await supertest(app.getHttpServer())
-      .post('/dynamic-ctx-handler')
-      .send({ name: 'Alice' })
-
-    expect(res.status).toBe(200)
-    expect(res.body).toEqual({ greeting: 'Hello, Alice', userId: 'dynamic-user-456' })
-
-    await app.close()
-  })
-
-  it('should read x-tenant-id request header and expose it via context factory', async () => {
-    // Real-world scenario: multi-tenant app where each request carries an
-    // x-tenant-id header. The context factory reads it once so every handler
-    // can access `context.tenantId` without touching the HTTP layer directly.
-    const contract = oc
-      .meta(openapi({ path: '/whoami' }))
-      .output(z.object({ tenantId: z.string() }))
-
-    @Controller()
-    class WhoAmIController {
-      @Implement(contract)
-      whoami() {
-        return implement(contract).handler(({ context }) => ({
-          tenantId: (context as any).tenantId,
-        }))
-      }
-    }
-
-    const moduleRef = await Test.createTestingModule({
-      controllers: [WhoAmIController],
-      imports: [
-        ORPCModule.forRoot({
-          context: (ctx) => {
-            const req = ctx.switchToHttp().getRequest()
-            return { tenantId: req.headers['x-tenant-id'] ?? 'default' } as any
-          },
-        }),
-      ],
-    }).compile()
-
-    const app = moduleRef.createNestApplication()
-    await app.init()
-
-    const res = await supertest(app.getHttpServer())
-      .post('/whoami')
-      .set('x-tenant-id', 'acme-corp')
-
-    expect(res.status).toBe(200)
-    expect(res.body).toEqual({ tenantId: 'acme-corp' })
-
-    await app.close()
-  })
-
-  it('should support context factory when using forRootAsync', async () => {
-    const contract = oc
-      .meta(openapi({ path: '/tenant-async' }))
-      .output(z.object({ tenantId: z.string() }))
-
-    @Controller()
-    class TenantController {
-      @Implement(contract)
-      tenant() {
-        return implement(contract).handler(({ context }) => ({
-          tenantId: (context as any).tenantId,
-        }))
-      }
-    }
-
-    const moduleRef = await Test.createTestingModule({
-      controllers: [TenantController],
-      imports: [
-        ORPCModule.forRootAsync({
-          useFactory: () => ({
-            context: (ctx: any) => {
-              const req = ctx.switchToHttp().getRequest()
-              return { tenantId: req.headers['x-tenant-id'] ?? 'default' } as any
-            },
-          }),
-        }),
-      ],
-    }).compile()
-
-    const app = moduleRef.createNestApplication()
-    await app.init()
-
-    const res = await supertest(app.getHttpServer())
-      .post('/tenant-async')
-      .set('x-tenant-id', 'acme-corp')
-
-    expect(res.status).toBe(200)
-    expect(res.body).toEqual({ tenantId: 'acme-corp' })
-
-    await app.close()
-  })
-})
