+++ packages/nest/src/implement.ts
@@ -6,7 +6,7 @@ import type { StandardBodyHint } from '@standardserver/core'
 import type { Request as ExpressRequest, Response as ExpressResponse } from 'express'
 import type { FastifyReply, FastifyRequest } from 'fastify'
 import type { Observable } from 'rxjs'
-import type { ContextFactory, NestStandardLazyRequest, ORPCModuleConfig } from './module'
+import type { NestStandardLazyRequest, ORPCModuleConfig } from './module'
 import { Readable } from 'node:stream'
 import { applyDecorators, Delete, Get, Head, HttpCode, HttpException, Inject, Injectable, Optional, Options, Patch, Post, Put, StreamableFile, UseInterceptors } from '@nestjs/common'
 import { HttpAdapterHost } from '@nestjs/core'
@@ -15,7 +15,7 @@ import { DEFAULT_OPENAPI_METHOD, getDynamicPathParams, getOpenAPIMeta } from '@o
 import { OpenAPIHandlerCodecCore } from '@orpc/openapi/standard'
 import { DEFAULT_SUCCESS_STATUS, getRouter, Procedure, unlazy } from '@orpc/server'
 import { StandardHandler } from '@orpc/server/standard'
-import { get, isAsyncIteratorObject, mergeHttpPath, stringifyJSON } from '@orpc/shared'
+import { get, isAsyncIteratorObject, mergeHttpPath, stringifyJSON, value } from '@orpc/shared'
 import { flattenStandardHeader, generateContentDisposition } from '@standardserver/core'
 import { toEventStream, toStandardLazyRequest } from '@standardserver/node'
 import { mergeMap } from 'rxjs'
@@ -153,9 +153,7 @@ export class ImplementInterceptor implements NestInterceptor {
         }, this.config)
 
         const result = await handler.handle(standardRequest, {
-          context: typeof this.config.context === 'function'
-            ? await (this.config.context as ContextFactory)(ctx)
-            : (this.config.context ?? {} as DefaultInitialContext),
+          context: await value(this.config.context ?? {} as DefaultInitialContext, ctx),
         })
 
         if (!result.matched) {
